+++ lib/smartgen/renderers/erb.rb
@@ -6,8 +6,6 @@ module Smartgen
       #
       # It exposes +markup_file+ variable and its +metadata+ to the ERB layout.
       def render(layout, markup_file, metadata={})
-        metadata = Smartgen::ObjectHash.new(metadata)
-
         ::ERB.new(layout).result(binding)
       end
     end