+++ spec/lib/smartgen/renderers/erb_spec.rb
@@ -17,16 +17,16 @@ describe Smartgen::Renderer::ERB do
   
   it "should render the given layout with metadata variable" do
     layout = "<html><body><%= markup_file.contents %><div><%= metadata[:some_key] %></div></body></html>"
-    subject.render(layout, markup_file, :some_key => 'some_value').should == "<html><body>#{contents}<div>some_value</div></body></html>"
+    subject.render(layout, markup_file, Smartgen::ObjectHash.new(:some_key => 'some_value')).should == "<html><body>#{contents}<div>some_value</div></body></html>"
   end
   
   it "should render the given layout with metadata variable, using methods instead of accessing keys" do
     layout = "<html><body><%= markup_file.contents %><div><%= metadata.some_key %></div></body></html>"
-    subject.render(layout, markup_file, :some_key => 'some_value').should == "<html><body>#{contents}<div>some_value</div></body></html>"
+    subject.render(layout, markup_file, Smartgen::ObjectHash.new(:some_key => 'some_value')).should == "<html><body>#{contents}<div>some_value</div></body></html>"
   end
 
   it "should render the given layout with metadata variable, using nested methods instead of accessing keys" do
     layout = "<html><body><%= markup_file.contents %><div><%= metadata.nested_hash.some_key %></div></body></html>"
-    subject.render(layout, markup_file, :nested_hash => {:some_key => 'some_value'}).should == "<html><body>#{contents}<div>some_value</div></body></html>"
+    subject.render(layout, markup_file, Smartgen::ObjectHash.new(:nested_hash => {:some_key => 'some_value'})).should == "<html><body>#{contents}<div>some_value</div></body></html>"
   end
 end